Revert "[Commit] Iteration 437 #errors = 22 <= 22."
This reverts commit 818ecc726bb28350f8cca21625962d634cba0d6f.

+++ MvcAppToModernize/WebApplication/Controllers/ProductController.cs
@@ -9,7 +9,6 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 
 namespace WebApplication.Controllers
 {
-using System;
     public class ProductViewModel
     {
         public IEnumerable<Product> Products { get; set; }
@@ -42,7 +41,7 @@ using System;
             var viewModel = new ProductViewModel
             {
                 Products = products,
-                CartItemCount = cartItems?.Sum(c => (System.Int32)c.Quantity) ?? 0
+                CartItemCount = cartItems?.Sum(c => c.Quantity) ?? 0
             };
 
             return Ok(viewModel);
